requirements.txt
APScheduler==3.10.4
astroid==3.2.4
backports.zoneinfo==0.2.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dill==0.3.9
et-xmlfile==1.1.0
flask==3.0.3
gunicorn==23.0.0
importlib-metadata==8.4.0
isort==5.13.2
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
numpy==1.24.4
openpyxl==3.1.5
packaging==24.1
pandas==2.0.3
platformdirs==4.3.6
pylint==3.2.7
python-dateutil==2.9.0.post0
pytz==2024.2
schedule==1.2.2
six==1.16.0
speedtest-cli==2.1.3
tomli==2.0.2
tomlkit==0.13.2
typing-extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
werkzeug==3.0.3
zipp==3.20.0
